requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
icalendar
python-dateutil
